Implementirana zaštita pristupa servletu gde je to potrebno

+++ SF-05-2018-OWP2019/src/cinema/TicketServlet.java
@@ -2,8 +2,6 @@ package cinema;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -13,27 +11,47 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import cinema.dao.MovieDAO;
 import cinema.dao.ProjectionDAO;
 import cinema.dao.SeatDAO;
 import cinema.dao.TicketDAO;
 import cinema.dao.UserDAO;
-import cinema.model.Movie;
 import cinema.model.Projection;
 import cinema.model.Seat;
 import cinema.model.Ticket;
 import cinema.model.User;
+import cinema.model.User.UserRole;
 
 @SuppressWarnings("serial")
 public class TicketServlet extends HttpServlet {
        
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
+			String loggedInUsername = (String) request.getSession().getAttribute("loggedInUsername");
+			if (loggedInUsername == null) {
+				request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+				return;
+			}
+			User loggedInUser = UserDAO.getUserByUsername(loggedInUsername);
+			if (loggedInUser == null) {
+				request.getRequestDispatcher("./LogoutServlet").forward(request, response);
+				return;
+			}
+			
+			if (!loggedInUser.isLoggedIn()) {
+				request.getRequestDispatcher("./LogoutServlet").forward(request, response);
+				return;
+			}
+			
 			String action = request.getParameter("action");
 			
 			switch (action) {
 				case "ticketID": {
 					Ticket ticket = TicketDAO.getTicketByID(Integer.parseInt(request.getParameter("id")));
+					
+					if (loggedInUser.getUserRole() != UserRole.ADMIN && !ticket.getUser().getUsername().equals(loggedInUsername)) {
+						request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+						return;
+					}
 					
 					Map<String, Object> data = new LinkedHashMap<String, Object>();
 					data.put("ticket", ticket);
@@ -47,6 +65,11 @@ public class TicketServlet extends HttpServlet {
 					break;
 				}
 				case "projectionID": {
+					if (loggedInUser.getUserRole() != UserRole.ADMIN) {
+						request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+						return;
+					}
+					
 					int projectionID = Integer.parseInt(request.getParameter("projectionID"));
 					List<Ticket> tickets = TicketDAO.getByProjectionID(projectionID);
 					Map<String, Object> data = new LinkedHashMap<String, Object>();
@@ -56,7 +79,13 @@ public class TicketServlet extends HttpServlet {
 					break; 
 				}
 				case "user": {
+					if (loggedInUser.getUserRole() != UserRole.ADMIN && !request.getParameter("username").equals(loggedInUsername)) {
+						request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+						return;
+					}
+					
 					List<Ticket> tickets = TicketDAO.getByUserUsername(request.getParameter("username"));
+					
 					Map<String, Object> data = new LinkedHashMap<String, Object>();
 					data.put("tickets", tickets);
 					
@@ -74,14 +103,35 @@ public class TicketServlet extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		try {
+			String loggedInUsername = (String) request.getSession().getAttribute("loggedInUsername");
+			if (loggedInUsername == null) {
+				request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+				return;
+			}
+			User loggedInUser = UserDAO.getUserByUsername(loggedInUsername);
+			if (loggedInUser == null) {
+				request.getRequestDispatcher("./LogoutServlet").forward(request, response);
+				return;
+			}
+			
+			if (!loggedInUser.isLoggedIn()) {
+				request.getRequestDispatcher("./LogoutServlet").forward(request, response);
+				return;
+			}
+			
 			String action = request.getParameter("action");
 			
 			switch (action) {
 				case "add": {
+					if (loggedInUser.getUserRole() != UserRole.USER) {
+						request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+						return;
+					}
+					
 					Projection projection = ProjectionDAO.getProjectionByID(Integer.parseInt(request.getParameter("projectionID")));
 					String[] stringSeatNumbers = request.getParameterValues("seatNumbers[]");
 					int hallID = Integer.parseInt(request.getParameter("hallID"));
-					User user = UserDAO.getUserByUsername((String)request.getSession().getAttribute("loggedInUsername"));
+					User user = loggedInUser;
 					
 					for (String number : stringSeatNumbers) {
 						int seatNumber = Integer.parseInt(number);
@@ -95,6 +145,11 @@ public class TicketServlet extends HttpServlet {
 					break;
 				}
 				case "delete": {
+					if (loggedInUser.getUserRole() != UserRole.ADMIN) {
+						request.getRequestDispatcher("./UnauthorizedServlet").forward(request, response);
+						return;
+					}
+					
 					Ticket ticket = TicketDAO.getTicketByID((Integer.parseInt(request.getParameter("id"))));
 					if (ticket != null) {
 						if (!TicketDAO.deleteTicket(ticket))
